# docs/api-contracts/openapi-v1-baseline.yaml
...
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
...
paths:
  /api/v1/users/{userId}:
    get:
      tags:
      - Users
      summary: Get a user
      description: Returns a user only when it belongs to the requested tenant.
      operationId: getUser
      parameters:
      - name: userId
        in: path
        description: User UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Missing or invalid identifier
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    put:
      tags:
      - Users
      summary: Update a user
      description: Updates email and display name within the tenant scope.
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        description: User UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Email already exists in tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/users:
    get:
      tags:
      - Users
      summary: List users
      description: 'Lists users belonging only to the requested tenant. Stage 03A: paginated.'
      operationId: listUsers
      parameters:
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      - name: params
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/PageRequestParams'
      responses:
        '200':
          description: User list (paginated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '400':
          description: Missing or invalid tenantId / pagination
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags:
      - Users
      summary: Create a user
      description: Creates a tenant-scoped user. Default status is INVITED.
      operationId: createUser
      parameters:
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Tenant not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Email already exists in tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/users/{userId}/suspend:
    patch:
      tags:
      - Users
      summary: Suspend a user
      description: Sets the user status to SUSPENDED.
      operationId: suspendUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: User suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/users/{userId}/deactivate:
    patch:
      tags:
      - Users
      summary: Deactivate a user
      description: Sets the user status to INACTIVE.
      operationId: deactivateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/users/{userId}/archive:
    patch:
      tags:
      - Users
      summary: Archive a user
      description: Sets the user status to ARCHIVED.
      operationId: archiveUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: User archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/users/{userId}/activate:
    patch:
      tags:
      - Users
      summary: Activate a user
      description: Sets the user status to ACTIVE.
      operationId: activateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: User activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/access/roles/{roleId}/deactivate:
    patch:
      tags:
      - role-controller
      operationId: deactivate
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/roles/{roleId}/archive:
    patch:
      tags:
      - role-controller
      operationId: archive
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/roles/{roleId}/activate:
    patch:
      tags:
      - role-controller
      operationId: activate
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/capabilities/{capabilityId}/deactivate:
    patch:
      tags:
      - capability-controller
      operationId: deactivate_1
      parameters:
      - name: capabilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
  /api/v1/access/capabilities/{capabilityId}/activate:
    patch:
      tags:
      - capability-controller
      operationId: activate_1
      parameters:
      - name: capabilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
...
components:
  schemas:
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
        displayName:
          type: string
          maxLength: 200
          minLength: 0
      required:
      - email
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - INVITED
          - SUSPENDED
          - ARCHIVED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiErrorResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldValidationError'
    FieldValidationError:
      type: object
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
...
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - ARCHIVED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CapabilityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
        displayName:
          type: string
          maxLength: 200
          minLength: 0
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - INVITED
          - SUSPENDED
          - ARCHIVED
      required:
      - email
...
    PageRequestParams:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortMetadata'
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items: {}
        page:
          $ref: '#/components/schemas/PageMetadata'
    SortMetadata:
      type: object
      properties:
        field:
          type: string
        direction:
          type: string
...